chatbot-service/pyproject.toml
[project]
name = "ai-agent-youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.18",
    "llama-index-readers-file>=0.6.0",
    "chromadb==1.5.5",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "pytest>=8.2.0",
    "requests>=2.32.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
